requirements.txt
# Core dependencies
pymupdf>=1.23.0
pillow>=10.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# AI/ML providers
openai>=1.0.0
anthropic>=0.18.0
